refactor(components): move IngredientCard to its own file
