+++ src/components/index.ts
@@ -4,3 +4,4 @@ export { default as ConsumableIdProgress } from "./ConsumableIdProgress";
 export { default as Consumables } from "./Consumables";
 export { default as ExploreByIngredient } from "./ExploreByIngredient";
 export { default as ExploreByNationality } from "./ExploreByNationality";
+export { default as IngredientCard } from "./IngredientCard";
